rename test files.
